requirements.txt
streamlit==1.37.0
streamlit-chat==0.1.1
streamlit-pdf-viewer==0.0.18.dev0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain-community==0.2.11
langchain-google-genai==1.0.10
faiss-cpu==1.8.0.post1